+++ src/test/java/sonia/scm/script/infrastructure/InitScriptContextListenerTest.java
@@ -32,6 +32,7 @@ import org.mockito.Captor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import sonia.scm.event.ScmEventBus;
 import sonia.scm.script.domain.ExecutionContext;
 import sonia.scm.script.domain.Executor;
 import sonia.scm.script.domain.InitScript;
@@ -56,6 +57,9 @@ class InitScriptContextListenerTest {
 
   @Mock
   private Executor executor;
+
+  @Mock
+  private ScmEventBus eventBus;
 
   @InjectMocks
   private InitScriptContextListener listener;
@@ -84,6 +88,7 @@ class InitScriptContextListenerTest {
 
     verify(executor, times(2)).execute(scriptCaptor.capture(), any(ExecutionContext.class));
     assertThat(scriptCaptor.getAllValues()).containsOnly(one, two);
+    verify(eventBus).post(any(ScmStartedEvent.class));
   }
 
   private InitScript createScript(String one) {